requirements.txt
llama-index-core==0.10.57
llama-index-llms-openai==0.1.29
llama-index-llms-openai-like==0.1.3
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-qdrant==0.2.8
fastapi==0.109.0
uvicorn[standard]==0.23.2
pydantic==1.10.14
python-dotenv==1.0.1
requests==2.31.0
pdf2image==1.16.3
Pillow==10.3.0
qdrant-client==1.11.0
transformers==4.40.2
sentence-transformers==2.7.0
openai==1.51.0
